SSL export fixes (from Adam Langley) [original from 1.0.1]

+++ ssl/s3_lib.c
@@ -2904,8 +2904,8 @@ SSL3_ENC_METHOD SSLv3_enc_data={
 	SSL3_MD_CLIENT_FINISHED_CONST,4,
 	SSL3_MD_SERVER_FINISHED_CONST,4,
 	ssl3_alert_code,
-	(int (*)(SSL *, unsigned char *, unsigned int, const char *,
-		 unsigned int, const unsigned char *, unsigned int,
+	(int (*)(SSL *, unsigned char *, size_t, const char *,
+		 size_t, const unsigned char *, size_t,
 		 int use_context))ssl_undefined_function,
 	};
